Added more dummy contents for visualization
Implemented some dummy book card to arrange the layout of the page.
Began working on implementing a modal dialog button to open and close,
I am getting stuck a trying to close the modal dialog when clicking
outside of it, which I've found a way but it is creating the scrollbar and
that is ruining the visual presentation.