Added visual changes and create book function
Layout and elements change for dialog modal pop up, as well as fixing
the header layout to look more aesthetic. Implemented a working function
to create new Book object when the submit button is clicked on the form,
reading out the form values for each input as parameters for the Book
constructor. Then initialized a display book function to create new
elements on the DOM base on each Book object attribute.